Moves lightning and convection plugins from nowcast to nowcasting.
